Made max_iterations a u32 to allow for deeper zooms

+++ fractals/fractal_utils/src/lib.rs
@@ -7,7 +7,7 @@ struct Color {
 }
 
 #[wasm_bindgen]
-pub fn mandelbrot(width: u32, height: u32, x_0: f64, x_1: f64, y_0: f64, y_1: f64, max_iterations: u8, colors: Vec<u8>) -> Vec<u8> {
+pub fn mandelbrot(width: u32, height: u32, x_0: f64, x_1: f64, y_0: f64, y_1: f64, max_iterations: u32, colors: Vec<u8>) -> Vec<u8> {
     println!("Computing fractal using width {} and height {}", width, height);
 
     //////// COLOR SCHEME ////////
@@ -20,7 +20,7 @@ pub fn mandelbrot(width: u32, height: u32, x_0: f64, x_1: f64, y_0: f64, y_1: f6
 
     let mut color_palette: Vec<Color> = Vec::new();
     let num_colors = color_list.len();
-    let section_width = max_iterations / (num_colors - 1) as u8;
+    let section_width = max_iterations / (num_colors - 1) as u32;
 
     let mut start_color: &Color;
     let mut end_color: &Color = &color_list[0]; //Initialized here in case num_colors = 1
@@ -64,7 +64,7 @@ pub fn mandelbrot(width: u32, height: u32, x_0: f64, x_1: f64, y_0: f64, y_1: f6
     println!("Using pixel width {} and pixel height {}", pixel_width, pixel_height);
 
     //Formula values
-    let (mut x, mut y, mut x_squared, mut y_squared, mut iterations): (f64, f64, f64, f64, u8);
+    let (mut x, mut y, mut x_squared, mut y_squared, mut iterations): (f64, f64, f64, f64, u32);
     let mut pixel_x = x_0;
     let mut pixel_y = y_0;
     let mut pixel_idx: usize = 0; //Keeps track of our location in the array
